+++ src/syllable.ts
@@ -9,21 +9,28 @@ import { Word } from "./word";
 interface SyllableCharToNameMap extends CharToNameMap {
   /* eslint-disable  @typescript-eslint/naming-convention */
   "\u{05B0}": "SHEVA"; // HEBREW POINT HATAF SHEVA (U+05B0)
+  /**
+   * Unlike a holam vav construction which has the holam present, the shureq has no vowel character.
+   */
+  "\u{05D5}\u{05BC}": "SHUREQ"; // HEBREW LETTER VAV (U+05D5) + HEBREW POINT DAGESH OR MAPIQ (U+05BC)
 }
 
 const sylCharToNameMap: SyllableCharToNameMap = {
   ...charToNameMap,
-  "\u{05B0}": "SHEVA"
+  "\u{05B0}": "SHEVA",
+  "\u{05D5}\u{05BC}": "SHUREQ"
 };
 
 interface SyllableNameToCharMap extends NameToCharMap {
   /* eslint-disable  @typescript-eslint/naming-convention */
   SHEVA: "\u{05B0}"; // HEBREW POINT HATAF SHEVA (U+05B0)
+  SHUREQ: "\u{05D5}\u{05BC}"; // HEBREW LETTER VAV (U+05D5) + HEBREW POINT DAGESH OR MAPIQ (U+05BC)
 }
 
 const sylNameToCharMap: SyllableNameToCharMap = {
   ...nameToCharMap,
-  SHEVA: "\u{05B0}"
+  SHEVA: "\u{05B0}",
+  SHUREQ: "\u{05D5}\u{05BC}"
 };
 
 /**
@@ -35,6 +42,8 @@ export class Syllable extends Node<Syllable> {
   #isAccented: boolean;
   #isFinal: boolean;
   #word: Word | null = null;
+  #cachedParts: SyllablePart[] | null = null;
+  #cachedStructure: [Consonant[], Vowel[], Consonant[]] | null = null;
 
   /**
    *
@@ -105,6 +114,10 @@ export class Syllable extends Node<Syllable> {
     return this.clusters.map((cluster) => cluster.chars).flat();
   }
 
+  private isVowelKeyOfSyllableCharToNameMap(vowel: string): vowel is keyof SyllableCharToNameMap {
+    return vowel in sylCharToNameMap;
+  }
+
   /**
    * Returns the vowel character of the syllable
    *
@@ -118,10 +131,26 @@ export class Syllable extends Node<Syllable> {
    * text.syllables[1].vowel;
    * // "\u{05B0}"
    * ```
+   *
+   * @description
+   * This returns a single vowel character, even for most mater lectionis (e.g. a holam vav would return the holam, not the vav).
+   * The only exception is a shureq, which returns the vav and the dagesh because there is no vowel character for a shureq.
    */
   get vowel(): keyof SyllableCharToNameMap | null {
+    const nucleus = this.nucleus.map((p) => p.text).join("");
+
+    // for regular vowel characters and shureqs, this should match
+    if (this.isVowelKeyOfSyllableCharToNameMap(nucleus)) {
+      return nucleus;
+    }
+
+    // for maters or text with mixed scripts (e.g. Hebrew and Latin), we have to extract the vowel character
     const match = this.text.match(vowelsCaptureGroupWithSheva);
-    return match ? (match[0] as keyof SyllableCharToNameMap) : match;
+    if (match && this.isVowelKeyOfSyllableCharToNameMap(match[0])) {
+      return match[0];
+    }
+
+    return null;
   }
 
   /**
@@ -136,6 +165,10 @@ export class Syllable extends Node<Syllable> {
    * // "PATAH"
    * text.syllables[1].vowelName;
    * // "SHEVA"
+   *
+   * @description
+   * This returns the vowel name, even for most mater lectionis (e.g. a holam vav would return the HOLAM, not the vav).
+   * The only exception is a shureq, which returns "SHUREQ" because there is no vowel character for a shureq.
    * ```
    */
   get vowelName(): SyllableCharToNameMap[keyof SyllableCharToNameMap] | null {
@@ -163,9 +196,21 @@ export class Syllable extends Node<Syllable> {
    * text.syllables[2].hasVowelName("SHEVA");
    * // false
    * ```
+   *
+   * @description
+   * This returns a boolean if the vowel character is present, even for most mater lectionis (e.g. in a holam vav construction, "HOLAM" would return true)
+   * The only exception is a shureq, because there is no vowel character for a shureq.
    */
   hasVowelName(name: keyof SyllableNameToCharMap): boolean {
-    if (!sylNameToCharMap[name]) throw new Error(`${name} is not a valid value`);
+    if (!sylNameToCharMap[name]) {
+      throw new Error(`${name} is not a valid value`);
+    }
+
+    if (name === "SHUREQ") {
+      // if any cluster has a shureq, then that should be the defacto vowel
+      return this.clusters.filter((c) => c.isShureq).length ? true : false;
+    }
+
     const isShevaSilent = name === "SHEVA" && this.clusters.filter((c) => c.hasVowel).length ? true : false;
     return !isShevaSilent && this.text.indexOf(sylNameToCharMap[name]) !== -1 ? true : false;
   }
@@ -259,6 +304,9 @@ export class Syllable extends Node<Syllable> {
    * ```
    */
   get parts(): SyllablePart[] {
+    if (this.#cachedParts) {
+      return this.#cachedParts;
+    }
     const parts: SyllablePart[] = [];
     let seenVowel = false;
     for (let i = 0; i < this.clusters.length; i++) {
@@ -376,6 +424,7 @@ export class Syllable extends Node<Syllable> {
       part.syllable = this;
       parts.push(part);
     }
+    this.#cachedParts = parts;
     return parts;
   }
 
@@ -397,6 +446,9 @@ export class Syllable extends Node<Syllable> {
    * ```
    */
   get structure(): [Consonant[], Vowel[], Consonant[]] {
+    if (this.#cachedStructure) {
+      return this.#cachedStructure;
+    }
     const onset: Consonant[] = [];
     const nucleus: Vowel[] = [];
     const coda: Consonant[] = [];
@@ -416,6 +468,7 @@ export class Syllable extends Node<Syllable> {
         }
       }
     }
+    this.#cachedStructure = [onset, nucleus, coda];
     return [onset, nucleus, coda];
   }