+++ test/word.test.ts
@@ -50,3 +50,28 @@ describe.each`
     });
   });
 });
+
+describe.each`
+  description      | heb              | isInConstructArray
+  ${"with maqqef"} | ${"בֶּן־אָדָ֕ם"} | ${[true, false]}
+  ${"now maqqef"}  | ${"בֶּן אָדָ֕ם"} | ${[false, false]}
+`("isInConstruct:", ({ description, heb, isInConstructArray }) => {
+  const text = new Text(heb);
+  test(`${description}`, () => {
+    expect(text.words.map((word) => word.isInConstruct)).toEqual(isInConstructArray);
+  });
+});
+
+describe("Implements node", () => {
+  const text = new Text("בֶּן־אָדָ֕ם");
+  const word = text.words[0];
+  test("prev", () => {
+    expect(word.prev).toEqual(null);
+  });
+  test("next", () => {
+    expect(word.next?.value?.text).toEqual("אָדָ֕ם");
+  });
+  test("value", () => {
+    expect(word.value?.text).toEqual("בֶּן־");
+  });
+});